Extract AICaptureModule from HackyAI
Another step towards making the AI more modular and HackyAI.cs less monolithic.
