+++ OpenRA.Mods.Common/OpenRA.Mods.Common.csproj
@@ -125,7 +125,7 @@
     <Compile Include="AI\BaseBuilder.cs" />
     <Compile Include="AI\HackyAI.cs" />
     <Compile Include="AI\AISupportPowerManager.cs" />
-    <Compile Include="AI\AICaptureManager.cs" />
+    <Compile Include="AI\AICaptureModule.cs" />
     <Compile Include="AI\AIHarvesterModule.cs" />
     <Compile Include="AI\Squad.cs" />
     <Compile Include="AI\StateMachine.cs" />
